refactor: RewardRange를 type 패키지로 이동
#62
